Build s390x gomplate in a Docker container
Gomplate is required for running the Instana agent in a Docker
container. There are no pre-built gomplate binaries for
Ubuntu 18.04 s390x. So build it from source with the script
"build_s390x_gomplate.sh" which creates a Docker image
"gomplate:latest", runs it as "gomplate", copies the gomplate
binary from the container to this project directory, and cleans up
container and image again.