s390x: Use and verify gomplate release 3.4.1
As a review result of #5, fetch the latest numbered release of
gomplate (v3.4.1) instead of using the master build. Also verify
a correct download of the source archive with sha256sum.
Use tar instead of unzip to get rid of a further dependency and
use the long option --version of gomplate for better readability.